dict_project/requirements.txt
asgiref==3.5.2
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.1.2
django-bootstrap3==22.2
mysqlclient==2.1.1
pdfminer.six==20220524
pycparser==2.21
sqlparse==0.4.3
tzdata==2022.4
